whisper_dialogue_toolkit/requirements.txt
faster-whisper==1.0.3
pydub==0.25.1
python-docx==1.1.2
soundfile==0.12.1
numpy==1.26.4
torch>=2.1.0